# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Asiscole — API del canal apoderado
  version: 0.1.0
  description: |
    Contrato de la API del canal de notificación de Asiscole.

    Es contract-first: ningún cliente consume un endpoint que no esté descrito aquí.

    ## Modelo de tokens (SRS 14.1)

    Hay dos tokens. El de sesión es el principal y prueba el login; el de datos es el
    secundario y autoriza las consultas, con vida corta.

    - `session_token`: 10 días. Se renueva por actividad entre el día 7 y el día 10.
      Solo sirve para `/auth/*`.
    - `data_token`: 15 a 60 minutos. Es el que llevan todos los endpoints de negocio.
      Si caduca con la sesión viva, el cliente llama a `/auth/refresh-data` de forma
      transparente y el usuario no percibe nada.

    Si `session_token` caduca o fue revocado (cierre de sesión, suspensión, transferencia
    aprobada en otro dispositivo o cierre forzado por el administrador), `refresh-data`
    responde 401 o 410 y el cliente vuelve al login.

    ## Identificación del apoderado

    El login usa el teléfono del apoderado en formato E.164 y el documento del estudiante.
    En la base de datos del colegio ese documento es la columna `estudiantes.codigo_barras`.

    ## Sesión única

    Solo hay una sesión activa por cuenta. Un segundo intento se deniega con
    `409 SESSION_ALREADY_ACTIVE`; nunca reemplaza a la sesión vigente. El dispositivo nuevo
    puede pedir una transferencia, que el dispositivo activo aprueba o rechaza.
servers:
...
security:
  - dataToken: []
paths:
...
  /auth/renew-session:
    post:
      tags: [auth]
      summary: Renovar el token de sesión (RF-A07)
      description: |
        Solo procede dentro de la ventana de renovación, entre el día 7 y el día 10 de
        vida del token. Fuera de esa ventana devuelve el token actual sin rotarlo.
      security:
        - sessionToken: []
      responses:
        "200":
          description: Sesión renovada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionCreada"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccountSuspended" }
        "410": { $ref: "#/components/responses/SessionExpired" }
...
components:
  securitySchemes:
    dataToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de datos, de vida corta. Es el que usan los endpoints de negocio.
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de sesión, de 10 días. Solo se usa en los endpoints de `/auth`.
    ingestApiKey:
      type: apiKey
      in: header
      name: X-Asiscole-Ingest-Key
      description: |
        Clave compartida entre el sistema escolar (Asiscole) y el canal.
        Configurar en el backend como `INGEST_API_KEY`.
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Código estable del catálogo. El cliente reacciona a esto, no al texto.
        message:
          type: string
          description: Mensaje en español, apto para mostrarse al usuario.
        request_id:
          type: string
      example:
        code: SESSION_ALREADY_ACTIVE
        message: Ya tienes una sesión activa en otro dispositivo.
        request_id: 6f1e2c3a-7b4d-4e2f-9a1c-8d5b3e7f0a21
    SesionCreada:
      type: object
      properties:
        session_token: { type: string }
        session_expira_en: { type: string, format: date-time }
        data_token: { type: string }
        data_expira_en: { type: string, format: date-time }
        perfil: { $ref: "#/components/schemas/Perfil" }
...
    Perfil:
      type: object
      properties:
        alias: { type: string, nullable: true }
        telefono: { type: string, description: Enmascarado, por ejemplo +51*****321 }
        estado:
          type: string
          enum: [activo, suspendido]
        motivo_suspension: { type: string, nullable: true }
        estudiante_activo_id: { type: integer, nullable: true }
        terminos_version:
          type: string
          nullable: true
          description: Versión de T&C aceptada (null si aún no aceptó en esta cuenta)
        terminos_aceptados_en:
          type: string
          format: date-time
          nullable: true
          description: Momento en que aceptó la versión vigente
...
  responses:
...
    Unauthenticated:
      description: 401 UNAUTHENTICATED — falta el token o es inválido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AccountSuspended:
      description: 403 ACCOUNT_SUSPENDED — la cuenta está suspendida
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    SessionExpired:
      description: 410 SESSION_EXPIRED — la sesión venció, hay que iniciar sesión otra vez
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...